tests: Refactor test class to utilize a repository test case
In order to eliminate redundancy across the various repository test classes,
such as when testing methods that the respective repositories inherit from the
BaseRepository class (e.g., list()), I added an abstract repository test case
that the various repository test classes should extend.

Closes #11
